Move shift cascade vertex (#25)
* Move shift_cascade_vertex to model; clean up configs; remove float_precision compatibility

* Rename estimate_charge_distribution -> charge_distribution_type

* Update tests

* Add compatibility note

---------
