+++ egenerator/model/source/track/stochastic_segment.py
@@ -47,12 +47,6 @@ class StochasticTrackSegmentModel(Source):
         """
         self.assert_configured(False)
 
-        # backwards compatibility for models that didn't define precision
-        if "float_precision" in config:
-            float_precision = config["float_precision"]
-        else:
-            float_precision = "float32"
-
         # -------------------------------------------
         # Define input parameters of track hypothesis
         # -------------------------------------------
@@ -86,7 +80,7 @@ class StochasticTrackSegmentModel(Source):
         if config["num_local_vars"] > 0:
             self._untracked_data["local_vars"] = new_weights(
                 shape=[1, 86, 60, config["num_local_vars"]],
-                float_precision=float_precision,
+                float_precision=config["float_precision"],
                 name="local_dom_input_variables",
             )
             num_inputs += config["num_local_vars"]
@@ -111,7 +105,7 @@ class StochasticTrackSegmentModel(Source):
             dilation_rate_list=None,
             hex_num_rotations_list=1,
             method_list=config["method_list"],
-            float_precision=float_precision,
+            float_precision=config["float_precision"],
         )
 
         return parameter_names
@@ -559,12 +553,18 @@ class StochasticTrackSegmentModel(Source):
         # -------------------------------------------
         # Get expected charge at DOM
         # -------------------------------------------
-        if config["estimate_charge_distribution"] is True:
+        if config["charge_distribution_type"] == "asymmetric_gaussian":
             n_charge = 3
-        elif config["estimate_charge_distribution"] == "negative_binomial":
+        elif config["charge_distribution_type"] == "negative_binomial":
             n_charge = 2
-        else:
+        elif config["charge_distribution_type"] == "poisson":
             n_charge = 1
+        else:
+            raise ValueError(
+                "Unknown charge distribution type: {}".format(
+                    config["charge_distribution_type"]
+                )
+            )
 
         # the result of the convolution layers are the latent variables
         dom_charges_trafo = tf.expand_dims(
@@ -602,7 +602,7 @@ class StochasticTrackSegmentModel(Source):
         # -------------------------------------
         # get charge distribution uncertainties
         # -------------------------------------
-        if config["estimate_charge_distribution"] is True:
+        if config["charge_distribution_type"] == "asymmetric_gaussian":
             sigma_scale_trafo = tf.expand_dims(
                 conv_hex3d_layers[-1][..., 1], axis=-1
             )
@@ -682,7 +682,7 @@ class StochasticTrackSegmentModel(Source):
             tensor_dict["dom_charges_variance"] = dom_charges_unc**2
             tensor_dict["dom_charges_log_pdf_values"] = dom_charges_llh
 
-        elif config["estimate_charge_distribution"] == "negative_binomial":
+        elif config["charge_distribution_type"] == "negative_binomial":
             """
             Use negative binomial PDF instead of Poisson to account for
             over-dispersion induces by systematic variations.
@@ -731,10 +731,17 @@ class StochasticTrackSegmentModel(Source):
             tensor_dict["dom_charges_variance"] = dom_charges_variance
             tensor_dict["dom_charges_log_pdf_values"] = dom_charges_llh
 
-        else:
+        elif config["charge_distribution_type"] == "poisson":
             # Poisson Distribution: variance is equal to expected charge
             tensor_dict["dom_charges_unc"] = tf.sqrt(dom_charges)
             tensor_dict["dom_charges_variance"] = dom_charges
+
+        else:
+            raise ValueError(
+                "Unknown charge distribution type: {}".format(
+                    config["charge_distribution_type"]
+                )
+            )
 
         # -------------------------------------------
         # Get times at which to evaluate DOM PDF
