+++ test/data/handler/test_modular.py
@@ -33,7 +33,6 @@ class TestModuleDataHandler(unittest.TestCase):
                 # settings for the label module
                 "label_module": "cascades.CascadeGeneratorLabelModule",
                 "label_settings": {
-                    "shift_cascade_vertex": False,
                     # logarithm on labels
                     # (x, y, z, zenith, azimuth, energy, time)?
                     "trafo_log": [
@@ -237,7 +236,6 @@ class TestModuleDataHandlerOnTestData(unittest.TestCase):
             # settings for the label module
             "label_module": "cascades.CascadeGeneratorLabelModule",
             "label_settings": {
-                "shift_cascade_vertex": False,
                 # logarithm on labels (x, y, z, zenith, azimuth, energy, time)?
                 "trafo_log": [False, False, False, False, False, True, False],
                 "label_key": "LabelsDeepLearning",
